decree_gen/requirements.txt
DAWG-Python==0.7.2
deprecation==2.1.0
docopt==0.6.2
img2pdf==0.4.4
loguru==0.6.0
lxml==4.9.1
packaging==21.3
pdf2jpg==1.1
pikepdf==6.2.4
Pillow==9.3.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==3.0.9
PyPDF2==2.11.2
python-docx==0.8.11
pytrovich==0.0.2
